test: test fixture 패키지 생성
