+++ src/test/java/fixture/PersonFixtures.java
@@ -1,0 +1,13 @@
+package fixture;
+
+public class PersonFixtures {
+    public static PersonV3 createPerson() {
+        return new PersonV3(
+                1L,
+                "yohan",
+                31,
+                "yohan@google.com",
+                1
+        );
+    }
+}
