+++ src/test/java/fixture/PersonV1.java
@@ -1,10 +1,10 @@
-package persistence.sql.ddl;
+package fixture;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 
 @Entity
-public class Person {
+public class PersonV1 {
     @Id
     private Long id;
     private String name;
