+++ src/test/java/fixture/PersonV3.java
@@ -1,0 +1,32 @@
+package fixture;
+
+import jakarta.persistence.*;
+
+@Table(name = "users")
+@Entity
+public class PersonV3 {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "nick_name")
+    private String name;
+
+    @Column(name = "old", length = 3)
+    private Integer age;
+
+    @Column(nullable = false)
+    private String email;
+
+    @Transient
+    private Integer index;
+
+    public PersonV3(Long id, String name, Integer age, String email, Integer index) {
+        this.id = id;
+        this.name = name;
+        this.age = age;
+        this.email = email;
+        this.index = index;
+    }
+}
