--- src/test/java/persistence/sql/ddl/PersonV2.java
@@ -1,19 +1,0 @@
-package persistence.sql.ddl;
-
-import jakarta.persistence.*;
-
-@Entity
-public class PersonV2 {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "nick_name")
-    private String name;
-
-    @Column(name = "old", length = 3)
-    private Integer age;
-
-    @Column(nullable = false)
-    private String email;
-}
